Enum multihash::Hash

pub enum Hash {
    SHA1,
    SHA2256,
    SHA2512,
    SHA3512,
    SHA3384,
    SHA3256,
    SHA3224,
    Keccak224,
    Keccak256,
    Keccak384,
    Keccak512,
    Blake2b,
    Blake2s,
}

List of types currently supported in the multihash spec.

Not all hash types are supported by this library.

Variants
SHA1

SHA-1 (20-byte hash size)

SHA2256

SHA-256 (32-byte hash size)

SHA2512

SHA-512 (64-byte hash size)

SHA3512

SHA3-512 (64-byte hash size)

SHA3384

SHA3-384 (48-byte hash size)

SHA3256

SHA3-256 (32-byte hash size)

SHA3224

SHA3-224 (28-byte hash size)

Keccak224

Keccak-224 (28-byte hash size)

Keccak256

Keccak-256 (32-byte hash size)

Keccak384

Keccak-384 (48-byte hash size)

Keccak512

Keccak-512 (64-byte hash size)

Blake2b

Encoding unsupported

Blake2s

Encoding unsupported

Methods

impl Hash

pub fn code(&self) -> u8

Get the corresponding hash code

pub fn size(&self) -> u8

Get the hash length in bytes

pub fn name(&self) -> &str

Get the human readable name

pub fn from_code(code: u8) -> Result<Hash, Error>