Trait multihash::MultihashDigest

pub trait MultihashDigest<T: TryFrom<u64>> {
    fn code(&self) -> T;

    fn digest(&self, data: &[u8]) -> MultihashGeneric<T>;

    fn input(&mut self, data: &[u8]);

    fn result(self) -> MultihashGeneric<T>;

    fn result_reset(&mut self) -> MultihashGeneric<T>;

    fn reset(&mut self);
}

The MultihashDigest trait specifies an interface common for all multihash functions.

Required methods

fn code(&self) -> T

The Mutlihash byte value.

fn digest(&self, data: &[u8]) -> MultihashGeneric<T>

Hash some input and return the digest.

Panics

Panics if the digest length is bigger than 2^32. This only happens for identity hasing.

fn input(&mut self, data: &[u8])

Digest data, updating the internal state.

This method can be called repeatedly for use with streaming messages.

Panics

Panics if the digest length is bigger than 2^32. This only happens for identity hashing.

fn result(self) -> MultihashGeneric<T>

Retrieve the computed MultihashGeneric, consuming the hasher.

fn result_reset(&mut self) -> MultihashGeneric<T>

Retrieve result and reset hasher instance.

This method sometimes can be more efficient compared to hasher re-creation.

fn reset(&mut self)

Reset hasher instance to its initial state.

Loading content...